2.2.0 Patch : optimization

+++ src/data/kaysaar/aotd/vok/campaign/econ/globalproduction/ui/components/optionpanels/FighterOptionPanelInterface.java
@@ -52,7 +52,10 @@ public class FighterOptionPanelInterface extends BaseOptionPanelManager implemen
         buttons.clear();
         sortingButtons.clear();
         chosenManu.clear();
-        searchbar.deleteAll();
+        if(searchbar!=null){
+            searchbar.deleteAll();
+        }
+
         this.mainPanel.removeComponent(panel);
     }
